backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
supabase==2.3.1
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0
pydantic==2.5.3
pydantic-settings==2.1.0
google-maps-services-python==4.10.0
